requirements.txt
astroid==2.0.4
colorama==0.3.9
img2pdf==0.3.1
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==4.2.5
mccabe==0.6.1
pgmagick==0.7.4
Pillow==5.2.0
pylint==2.1.1
PyPDF2==1.26.0
python-pptx==0.6.13
six==1.11.0
typed-ast==1.1.0
virtualenv==16.0.0
wrapt==1.10.11
XlsxWriter==1.1.0
